LayoutTests/fast/selectors/querySelector-simple-child-backtracking-expected.txt

Test backtracking with a single child selector relation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("ul>li p").length is 2
PASS document.querySelectorAll("ul>li p")[0].id is "target1"
PASS document.querySelectorAll("ul>li p")[1].id is "target2"
PASS successfullyParsed is true

TEST COMPLETE
